+++ server.js
@@ -1,6 +1,7 @@
 const express = require('express');
 const app = express();
-const cors = require('cors')
+const cors = require('cors');
+const corsOptions = require('./src/config/corsOptions.js');
 require('dotenv').config();
 
 const port = process.env.PORT;
@@ -9,17 +10,6 @@ const budget = require('./src/routes/budget');
 const expenses = require('./src/routes/expenses');
 const category = require('./src/routes/category');
 
-const whitelist = ['http://localhost:8080', 'http://localhost:3002']
-const corsOptions = {
-    origin: (origin, callback) => {
-        if(whitelist.indexOf(origin) !== -1 || !origin) {
-            callback(null, true)
-        } else {
-            callback(new Error('Not allowed by CORS'))
-        }   
-    },
-    optionsSuccessStatus: 200
-}
 app.use(cors(corsOptions));
 
 app.use(express.json());
