+++ src/config/configDB.js
@@ -10,6 +10,4 @@ const pool = new Pool({
     // ssl: true
 });
 
-module.exports = {
-    pool
-};
+module.exports = { pool };
